requirements.txt
amqp==2.5.2
asgiref==3.2.7
astroid==2.3.3
billiard==3.6.3.0
bottle==0.12.18
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
crc16==0.1.1
Django==3.0.4
django-appconf==1.0.3
django-autoslug==1.9.6
django-compressor==2.4
django-debug-toolbar==2.2
django-filter==2.2.0
django-filters==0.2.1
django-fsm==2.7.0
django-haystack==3.0b1
django-ipware==2.1.0
django-mdeditor==0.1.17
django-uuslug==1.2.0
djangorestframework==3.11.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jieba==0.42.1
jsonpickle==1.3
kombu==4.6.8
lazy-object-proxy==1.4.3
markdown2==2.3.8
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.6
Pillow==7.0.0
psutil==5.7.0
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-logstash==0.4.6
python-memcached==1.59
python-redis==0.2.2
python-slugify==4.0.0
pytz==2019.3
raven==6.10.0
rcssmin==1.0.6
redis==3.4.1
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
WeRoBot==1.10.1
Whoosh==2.7.4
wrapt==1.11.2
xmltodict==0.12.0
zipp==3.1.0
